バフカードがプレイヤーの状態によって与える効果を調整させる機能実装途中

+++ Assets/Scripts/Card/AttackCard.cs
@@ -16,6 +16,10 @@ public class AttackCard : CardBase, IAttackCard
         m_tooltip.text = $"{m_cardData.m_cardData.Damage}ダメージを与える";
     }
 
+    /// <summary>
+    /// 与ダメージ計算。ダメージを受ける（Dropされる）側で受け取る必要がある
+    /// </summary>
+    /// <returns>計算後のダメージと与える状態</returns>
     public int[] GetDamage()
     {
         OnCast();
@@ -23,7 +27,7 @@ public class AttackCard : CardBase, IAttackCard
         nums[(int)BuffDebuff.Damage] = m_cardData.m_cardData.Damage;
         for (int i = 0; i < m_cardData.m_cardData.GiveStateNum; i++)
         {
-            nums[(int)m_cardData.m_cardData.GiveState(i)] = (int)m_cardData.m_cardData.GiveStateTrun(i);
+            nums[(int)m_cardData.m_cardData.GiveState(i)] = m_cardData.m_cardData.GiveStateTrun(i);
         }
         return nums;
     }
